feat: rename files in `module-manager` to match kebab-case and create missing index.ts in /lib
