+++ packages/module-manager/lib/recover-feature.ts
@@ -1,5 +1,5 @@
 import { DEFAULT_CHOICES, RECOVER_CHOICE_QUESTION } from './const.js';
-import { recoverModule } from './modulesHandler.js';
+import { recoverModule } from './modules-handler.js';
 import { promptSelection } from './utils.js';
 import { existsSync, readdirSync } from 'node:fs';
 import { resolve } from 'node:path';
